requirements.txt
asgiref==3.8.1
Django==5.1.2
mysqlclient==2.2.4
sqlparse==0.5.1
django-extensions==3.2.3
djangorestframework==3.15.0
qrcode==7.4.2 
Pillow==10.1.0
requests==2.32.3
python-dotenv==1.0.1
